requirements.txt
asttokens==3.0.0
boto3==1.39.3
botocore==1.39.3
certifi==2025.6.15
charset-normalizer==3.4.2
comm==0.2.2
debugpy==1.8.14
decorator==5.2.1
dnspython==2.7.0
executing==2.2.0
idna==3.10
ipykernel==6.29.5
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jmespath==1.0.1
jupyter_client==8.6.3
jupyter_core==5.8.1
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.3.1
packaging==25.0
pandas==2.3.0
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
Pygments==2.19.2
pymongo==4.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pyzmq==27.0.0
requests==2.31.0
s3transfer==0.13.0
six==1.17.0
stack-data==0.6.3
tornado==6.5.1
traitlets==5.14.3
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
